moved non-allocating routines to new file
